* emacs-24.4-mac-5.2 (2014-11-21)
** Fixed bugs

*** The function `mac-osa-script' does not process ARGS (6th and later
arguments) correctly.

** Improvements

*** Even if the current keyboard layout maps some key with the control
modifier to insertion of a character, which would be recognized as an
ordinary insertion in applications that use the standard Cocoa text
view, it is interpreted as a combination with the control modifier.
This behavior is consistent with the Mac port version 4.x and earlier,
and also obviates the previous fix for "C-_" on JIS keyboard.

*** If /usr/include does not exist, then the Xcode SDK dir is examined
as a candidate of libxml2 header location if it does not contain any
special characters or blanks.

*** The function `mac-osa-script' now accepts compiled code, which can
be generated by a new function `mac-osa-compile', and a file name as
well as a script source.
Note: As of OS X 10.10.1, specifying compiled code generated from a
JavaScript source causes crash.  I think this is a bug in OSAKit.

*** New function `mac-do-applescript', which is implemented on top of
`mac-osa-script' and intended as a future replacement of
`do-applescript'.

*** New functions and hooks for interaction with the Text Input Source
Services on Mac OS X 10.5 and later: `mac-input-source',
`mac-input-source-list', `mac-select-input-source',
`mac-deselect-input-source',
`mac-selected-keyboard-input-source-change-hook', and
`mac-enabled-keyboard-input-sources-change-hook'.

*** New global minor mode `mac-auto-ascii-mode' for Mac OS X 10.5 and
later.  If enabled, it automatically selects the most-recently-used
ASCII-capable keyboard input source on some occasions: after prefix
key (bound in the global keymap) press such as C-x and M-g, and at the
start of minibuffer input.
Note: Strictly speaking, its implementation has a timing issue: the
Lisp event queue may already have some input events that have been
processed by some previous keyboard input source but yet to be
processed by the Lisp interpreter.